fixed di related problems BREAKING
